feat: Add wnfs story (work in progress)

+++ packages/plugins/experimental/plugin-wnfs/src/blockstore.ts
@@ -8,7 +8,7 @@ import type { CID } from 'multiformats';
 
 import { storeName } from './common';
 
-export const create = (apiHost: string) => {
+export const create = (apiHost?: string) => {
   return new MixedBlockstore(apiHost);
 };
 
@@ -17,10 +17,10 @@ export const create = (apiHost: string) => {
  * and which uses an indexedDB blockstore as a client cache.
  */
 class MixedBlockstore extends BaseBlockstore {
-  readonly #apiHost: string;
+  readonly #apiHost: string | undefined;
   readonly #idbStore: IDBBlockstore;
 
-  constructor(apiHost: string) {
+  constructor(apiHost?: string) {
     super();
 
     this.#apiHost = apiHost;
@@ -31,17 +31,20 @@ class MixedBlockstore extends BaseBlockstore {
     await this.#idbStore.open();
   }
 
-  url(cid?: CID) {
+  url(apiHost: string, cid?: CID) {
     const path = cid ? cid.toString() : '';
-    return `${this.#apiHost}/api/file/${path}`;
+    return `${apiHost}/api/file/${path}`;
   }
 
   // Blockstore implementation
   override async delete(key: CID): Promise<void> {
     await this.#idbStore.delete(key);
-    await fetch(this.url(key), {
-      method: 'DELETE',
-    });
+
+    if (this.#apiHost) {
+      await fetch(this.url(this.#apiHost, key), {
+        method: 'DELETE',
+      });
+    }
   }
 
   override async get(key: CID): Promise<Uint8Array> {
@@ -49,11 +52,15 @@ class MixedBlockstore extends BaseBlockstore {
       return await this.#idbStore.get(key);
     }
 
-    return await fetch(this.url(key), {
-      method: 'GET',
-    })
-      .then((r) => r.arrayBuffer())
-      .then((r) => new Uint8Array(r));
+    if (this.#apiHost) {
+      return await fetch(this.url(this.#apiHost, key), {
+        method: 'GET',
+      })
+        .then((r) => r.arrayBuffer())
+        .then((r) => new Uint8Array(r));
+    }
+
+    return await this.#idbStore.get(key);
   }
 
   override async has(key: CID): Promise<boolean> {
@@ -61,17 +68,24 @@ class MixedBlockstore extends BaseBlockstore {
       return true;
     }
 
-    return await fetch(this.url(key), {
-      method: 'HEAD',
-    }).then((r) => r.ok);
+    if (this.#apiHost) {
+      return await fetch(this.url(this.#apiHost, key), {
+        method: 'HEAD',
+      }).then((r) => r.ok);
+    }
+
+    return false;
   }
 
   override async put(key: CID, val: Uint8Array): Promise<CID> {
     await this.#idbStore.put(key, val);
-    await fetch(this.url(key), {
-      method: 'POST',
-      body: val,
-    });
+
+    if (this.#apiHost) {
+      await fetch(this.url(this.#apiHost, key), {
+        method: 'POST',
+        body: val,
+      });
+    }
 
     return key;
   }
